requirements.txt
Django==1.5.4
argparse==1.2.1
distribute==0.6.34
dj-database-url==0.2.2
dj-static==0.0.5
django-admin-bootstrapped==0.4.3
django-ajax-selects==1.3.1
django-toolbelt==0.0.1
gunicorn==18.0
ipython==1.1.0
psycopg2==2.5.1
pytz==2013.7
reportlab==2.7
static==0.4
wsgiref==0.1.2
